Move `Observable#toPromise()` to `React\Promise::fromObservable()`
